report/requirements-report.txt
# Dokumentgenerering – flotte rapporter fra introduktion
# Installer: pip install -r report/requirements-report.txt

weasyprint>=62.0
reportlab>=4.0.0
fpdf2>=2.7.0
jinja2>=3.1.0
python-docx>=1.0.0
markdown>=3.5.0
